Reorganize javascript directories
